collector/requirements.txt
influxdb-client==1.38.0
PyYAML==6.0.1
requests==2.31.0
